GAP-130: identifiers minted as UUID v7, written as strings, shown as tags
D-56, D-60, D-61; DN-31 §5.1 amendment 1. DecisionId, PlanId and
PendingApprovalId hold a u128 minted with Uuid::now_v7 in the workflow and
the planner, replacing counters that restarted at 1 per process.
gungnir_model::identifier writes the hyphenated form and reads it or a
number, so old journals read; the handoff delivered to an effector and
published for exchange are no longer null. Panels and alerts show the short
tag, PN-07 and PN-20 the whole id with a copy control, logs and audit
entries the whole id. SCHEMA_VERSION is 4 and ApprovalRequested is removed.
The rationale names no throwaway plan id, and the delivery stub now checks
the body it receives.

+++ gungnir-app/src/rehearsal.rs
@@ -301,8 +301,10 @@ pub fn tick(state: &mut AppState) {
         crate::update::publish(state, now, Event::Rehearsal(event));
     }
     for p in plans {
+        // The seed's own number (`P-1183`), not a minted identifier: a task card names
+        // it, so it is shown as the seed gives it (D-61).
         let plan = PlanView::intercept(
-            PlanId(p.id),
+            PlanId(u128::from(p.id)),
             now,
             vec![InterceptSolutionView {
                 resource: ResourceId(p.resource),
